samples/image-classify/requirements.txt
pillow
pandas
exifread
fastmcp
